# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeadMind API",
    "description": "\nRead access to the LeadMind corpus, and the human review queue for duplicates the pipeline\ndeliberately refused to merge.\n\n**Data quality is not lead quality.** Every `quality` field in this API answers *how much do we\nreliably know about this record* \u2014 completeness, reachability, verifiability. Whether a lead is\nworth contacting is a separate judgement with different inputs, and arrives in Phase 5.\n\n**`unknown` is not `unreachable`.** A domain with no MX record is a measurement. A resolver\ntimeout is the absence of one. They are different values everywhere in this API and must not be\ncollapsed by a client.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List leads",
        "operationId": "list_leads_api_v1_leads_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on name or any identifier value",
              "title": "Q"
            },
            "description": "Substring match on name or any identifier value"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Controlled vertical slug; repeatable",
              "title": "Category"
            },
            "description": "Controlled vertical slug; repeatable"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Gazetteer location slug; repeatable",
              "title": "Location"
            },
            "description": "Gazetteer location slug; repeatable"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Indian state; repeatable",
              "title": "State"
            },
            "description": "Indian state; repeatable"
          },
          {
            "name": "entity_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityKind"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Kind"
            }
          },
          {
            "name": "has",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentifierKind"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Require these channels; repeatable and combined with AND",
              "title": "Has"
            },
            "description": "Require these channels; repeatable and combined with AND"
          },
          {
            "name": "missing",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentifierKind"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Require these channels to be absent",
              "title": "Missing"
            },
            "description": "Require these channels to be absent"
          },
          {
            "name": "min_quality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Quality"
            }
          },
          {
            "name": "max_quality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Quality"
            }
          },
          {
            "name": "min_followers",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Followers"
            }
          },
          {
            "name": "max_followers",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Followers"
            }
          },
          {
            "name": "owned_website",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Whether the lead has a domain of its own, as opposed to a social URL",
              "title": "Owned Website"
            },
            "description": "Whether the lead has a domain of its own, as opposed to a social URL"
          },
          {
            "name": "mailbox_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VerificationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Measured MX outcome. `unreachable` means proven undeliverable; `unknown` means the resolver failed or the check has not run \u2014 they are not the same thing and this filter does not conflate them.",
              "title": "Mailbox Status"
            },
            "description": "Measured MX outcome. `unreachable` means proven undeliverable; `unknown` means the resolver failed or the check has not run \u2014 they are not the same thing and this filter does not conflate them."
          },
          {
            "name": "website_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VerificationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Website Status"
            }
          },
          {
            "name": "mail_provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Mail host inferred from MX, e.g. `google`, `microsoft`",
              "title": "Mail Provider"
            },
            "description": "Mail host inferred from MX, e.g. `google`, `microsoft`"
          },
          {
            "name": "placeholder_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymised `Advertiser 13887200` names",
              "title": "Placeholder Name"
            },
            "description": "Anonymised `Advertiser 13887200` names"
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "multi_branch",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Leads whose company has more than one branch",
              "title": "Multi Branch"
            },
            "description": "Leads whose company has more than one branch"
          },
          {
            "name": "issue_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Leads carrying a given validation issue code",
              "title": "Issue Code"
            },
            "description": "Leads carrying a given validation issue code"
          },
          {
            "name": "include_merged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include leads a reviewer confirmed as duplicates. They are hidden by default and never deleted.",
              "default": false,
              "title": "Include Merged"
            },
            "description": "Include leads a reviewer confirmed as duplicates. They are hidden by default and never deleted."
          },
          {
            "name": "rubric_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Score against a specific rubric version instead of the current one",
              "title": "Rubric Version"
            },
            "description": "Score against a specific rubric version instead of the current one"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of quality, followers, name, created, updated. Prefix with `-` for descending. Every ordering is tiebroken on id so paging is stable.",
              "default": "-quality",
              "title": "Sort"
            },
            "description": "One of quality, followers, name, created, updated. Prefix with `-` for descending. Every ordering is tiebroken on id so paging is stable."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps.",
              "title": "Page Size"
            },
            "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps."
          }
        ],
        "responses": {
          "200": {
            "description": "One page of leads, newest rubric scores attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LeadSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "One lead, with its gaps",
        "operationId": "get_lead_api_v1_leads__lead_id__get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/quality": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Why this lead scored what it scored",
        "operationId": "get_lead_quality_api_v1_leads__lead_id__quality_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          },
          {
            "name": "rubric_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rubric Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadQualityDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/provenance": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "The spreadsheet rows this lead came from",
        "operationId": "get_lead_provenance_api_v1_leads__lead_id__provenance_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityKind": {
        "type": "string",
        "enum": [
          "person",
          "business",
          "unknown"
        ],
        "title": "EntityKind",
        "description": "Whether a lead's name denotes a person, a business, or cannot be told apart."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentifierKind": {
        "type": "string",
        "enum": [
          "email",
          "phone",
          "whatsapp",
          "website",
          "facebook",
          "instagram",
          "youtube",
          "linkedin"
        ],
        "title": "IdentifierKind",
        "description": "Channels a lead can be reached or identified by.\n\nModelled as rows in ``lead_identifiers`` rather than eight sparse columns on ``leads``:\nthe source data fills between 12% and 100% of them, dedup needs to query across all of\nthem, and a ninth channel should be a row, not a migration."
      },
...
      "LeadDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "entity_kind": {
            "$ref": "#/components/schemas/EntityKind"
          },
          "is_placeholder_name": {
            "type": "boolean",
            "title": "Is Placeholder Name"
          },
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "fb_category_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fb Category Raw"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/IdentifierKind"
            },
            "type": "array",
            "title": "Channels",
            "description": "Which contact channels exist for this lead"
          },
          "primary_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Email"
          },
          "primary_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Phone"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "followers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FollowersOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationOut"
          },
          "issue_count": {
            "type": "integer",
            "title": "Issue Count",
            "default": 0
          },
          "merged": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MergeInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when a reviewer confirmed this lead as a duplicate of another. The row is intact and hidden from listings, not deleted."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/IdentifierOut"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricOut"
            },
            "type": "array",
            "title": "Metrics"
          },
          "source_queries": {
            "items": {
              "$ref": "#/components/schemas/SourceQueryOut"
            },
            "type": "array",
            "title": "Source Queries"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueOut"
            },
            "type": "array",
            "title": "Issues",
            "description": "Recorded validation failures. Nothing was dropped because of them \u2014 they are the explicit gaps in this record."
          },
          "siblings": {
            "items": {
              "$ref": "#/components/schemas/SiblingOut"
            },
            "type": "array",
            "title": "Siblings",
            "description": "Other leads at the same company \u2014 branches, usually"
          },
          "duplicate_candidate_count": {
            "type": "integer",
            "title": "Duplicate Candidate Count",
            "default": 0
          },
          "niche_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche Raw"
          },
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "entity_kind",
          "is_placeholder_name",
          "created_at",
          "updated_at"
        ],
        "title": "LeadDetail",
        "description": "One lead, with everything known about it and everything not known about it."
      },
...
      "Page_LeadSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Rows matching the filter, ignoring pagination"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "description": "Total number of pages at this page size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "Page[LeadSummary]"
      },
      "ProvenanceOut": {
        "properties": {
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lead Id"
          },
          "source_records": {
            "items": {
              "$ref": "#/components/schemas/SourceRecordOut"
            },
            "type": "array",
            "title": "Source Records"
          }
        },
        "type": "object",
        "required": [
          "lead_id"
        ],
        "title": "ProvenanceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "verified",
          "unreachable",
          "unknown",
          "skipped"
        ],
        "title": "VerificationStatus",
        "description": "The outcome of a check.\n\nThe distinction that matters is between ``UNREACHABLE`` and ``UNKNOWN``. The first is a\nmeasurement: the domain resolved and has no mail exchanger. The second is an absence of\nmeasurement: the resolver timed out, the network was down, the check was never run. Folding\nthem together would let an outage silently downgrade thousands of good leads."
      }
    }
  }
}